runtime: enhance WebAssembly support with WebGPU and OffscreenCanvas integration
- Added a GPU broker worker to manage WebGPU operations, allowing efficient rendering in a WebAssembly context.
- Implemented OffscreenCanvas support for rendering, enabling the engine to utilize the browser's graphics capabilities.
- Updated the runtime to handle viewport dimensions and pixel density for better rendering quality.
- Enhanced threading model to support GPU calls across multiple workers, improving performance and responsiveness.
- Introduced a new test application demonstrating the engine's capabilities in a browser environment.

+++ stappler/data/SPDataValue.h
@@ -371,7 +371,7 @@ public:
 	void setBool(bool value) { *this = value; }
 	void setInteger(int32_t value) { *this = value; }
 	void setInteger(int64_t value) { *this = value; }
-	void setInteger(size_t value) { *this = value; }
+	void setInteger(size_t value) { *this = int64_t(value); }
 	void setDouble(float value) { *this = value; }
 	void setDouble(double value) { *this = value; }
 	void setString(const char *value) { *this = value; }
